backend/api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import PropertyViewSet, QueryViewSet

router = DefaultRouter()
router.register(r'properties', PropertyViewSet, basename='property')
router.register(r'queries', QueryViewSet, basename='query')

urlpatterns = [
    path('', include(router.urls)),
]
