backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
pandas>=2.1
openpyxl>=3.11
google-generativeai>=0.3
gunicorn>=21.2
requests>=2.31
